+++ bash/main.sh
@@ -11,10 +11,10 @@ source ${RECEIVER_BASH_DIRECTORY}/functions.sh
 adsb_decoder_installed="false"
 install_adsb_decoder="false"
 
-if [[ $(dpkg-query -W -f='${STATUS}' dump1090-fa 2>/dev/null | grep -c "ok installed") == 1 ]] ; then
+if [[ $(dpkg-query -W -f='${STATUS}' dump1090-fa 2>/dev/null | grep -c "ok installed") == 1 ]]; then
     adsb_decoder_installed="true"
     chosen_adsb_decoder="dump1090-fa"
-    if [[ $(sudo dpkg -s dump1090-fa 2>/dev/null | grep -c "Version: ${dump1090_fa_current_version}") == 0 ]] ; then
+    if [[ $(sudo dpkg -s dump1090-fa 2>/dev/null | grep -c "Version: ${dump1090_fa_current_version}") == 0 ]]; then
         whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
                  --title "FlightAware Dump1090 Upgrade Available" \
                  --defaultno \
@@ -34,6 +34,19 @@ if [[ $(dpkg-query -W -f='${STATUS}' dump1090-fa 2>/dev/null | grep -c "ok insta
     fi
 fi
 
+if [[ $(dpkg-query -W -f='${STATUS}' readsb 2>/dev/null | grep -c "ok installed") == 1 ]]; then
+    adsb_decoder_installed="true"
+    chosen_adsb_decoder="readsb"
+    whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
+             --title "Reinstall Readsb Decoder" \
+             --defaultno \
+             --yesno "The option to rebuild and reinstall Readsb is available.\n\nWould you like to rebuild and reinstall Readsb?" \
+             9 65
+    if [[ $? == 0 ]]; then
+        install_adsb_decoder="true"
+    fi
+fi
+
 if [[ "${adsb_decoder_installed}" == "false" ]]; then
     chosen_adsb_decoder=$(whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
                                    --title "ADS-B Decoder Selection" \
@@ -41,6 +54,7 @@ if [[ "${adsb_decoder_installed}" == "false" ]]; then
                                    16 100 9 \
                                    "None" "Do not install an ADS-B decoder." \
                                    "dump1090-fa" "FlightAware's version of the dump1090 decoder." \
+                                   "readsb" "Wiedehopf's detached fork of readsb." \
                                    3>&2 2>&1 1>&3)
     exit_status=$?
     if [[ $exit_status != 0 || "${chosen_uat_decoder}" == "None" ]]; then
@@ -51,6 +65,14 @@ fi
 function install_dump1090_fa() {
     chmod +x ${RECEIVER_BASH_DIRECTORY}/decoders/dump1090-fa.sh
     ${RECEIVER_BASH_DIRECTORY}/decoders/dump1090-fa.sh
+    if [[ $? != 0 ]] ; then
+        exit 1
+    fi
+}
+
+function install_readsb() {
+    chmod +x ${RECEIVER_BASH_DIRECTORY}/decoders/readsb.sh
+    ${RECEIVER_BASH_DIRECTORY}/decoders/readsb.sh
     if [[ $? != 0 ]] ; then
         exit 1
     fi
@@ -420,10 +442,8 @@ function install_beastsplitter() {
 
 # Duck DNS
 if [[ ! -f "${RECEIVER_BUILD_DIRECTORY}/duckdns/duck.sh" ]]; then
-    # Duck DNS does not appear to be set up on this device.
     extras_list=("${extras_list[@]}" 'Duck DNS Free Dynamic DNS Hosting' '' OFF)
 else
-    # Offer the option to install/setup Duck DNS once more.
     extras_list=("${extras_list[@]}" 'Duck DNS Free Dynamic DNS Hosting (reinstall)' '' OFF)
 fi
 
@@ -431,6 +451,36 @@ function install_duckdns() {
     chmod +x ${RECEIVER_BASH_DIRECTORY}/extras/duckdns.sh
     ${RECEIVER_BASH_DIRECTORY}/extras/duckdns.sh
     if [[ $? -ne 0 ]] ; then
+        exit 1
+    fi
+}
+
+# Graphs1090
+if [[ ! -f /lib/systemd/system/graphs1090.service ]]; then
+    extras_list=("${extras_list[@]}" 'Graphs1090' '' OFF)
+else
+    extras_list=("${extras_list[@]}" 'Graphs1090 (reinstall)' '' OFF)
+fi
+
+function install_graphs1090() {
+    chmod +x ${RECEIVER_BASH_DIRECTORY}/extras/graphs1090.sh
+    ${RECEIVER_BASH_DIRECTORY}/extras/graphs1090.sh
+    if [[ $? != 0 ]] ; then
+        exit 1
+    fi
+}
+
+# tar1090
+if [[ ! -f /lib/systemd/system/tar1090.service ]]; then
+    extras_list=("${extras_list[@]}" 'tar1090' '' OFF)
+else
+    extras_list=("${extras_list[@]}" 'tar1090 (reinstall)' '' OFF)
+fi
+
+function install_tar1090() {
+    chmod +x ${RECEIVER_BASH_DIRECTORY}/extras/tar1090.sh
+    ${RECEIVER_BASH_DIRECTORY}/extras/tar1090.sh
+    if [[ $? != 0 ]] ; then
         exit 1
     fi
 }
@@ -464,6 +514,9 @@ else
         case ${chosen_adsb_decoder} in
             "dump1090-fa")
                 confirmation_message="${confirmation_message}\n  * FlightAware dump1090"
+                ;;
+            "readsb")
+                confirmation_message="${confirmation_message}\n  * Readsb"
                 ;;
         esac
     fi
@@ -538,6 +591,9 @@ if [[ "${install_adsb_decoder}" == "true" ]]; then
     case ${chosen_adsb_decoder} in
         "dump1090-fa")
              install_dump1090_fa
+             ;;
+        "Readsb")
+             install_readsb
              ;;
     esac
 fi
@@ -658,6 +714,12 @@ if [[ -s "${RECEIVER_ROOT_DIRECTORY}/EXTRAS_CHOICES" ]]; then
             "Duck DNS Free Dynamic DNS Hosting"|"Duck DNS Free Dynamic DNS Hosting (reinstall)")
                 run_duckdns_script="true"
                 ;;
+            "Graphs1090"|"Graphs1090 (reinstall)")
+                run_graphs1090_script="true"
+                ;;
+            "tar1090"|"tar1090 (reinstall)")
+                run_tar1090_script="true"
+                ;;
         esac
     done < ${RECEIVER_ROOT_DIRECTORY}/EXTRAS_CHOICES
 fi
@@ -668,6 +730,14 @@ fi
 
 if [[ "${run_duckdns_script}" == "true" ]]; then
     install_duckdns
+fi
+
+if [[ "${run_graphs1090_script}" == "true" ]]; then
+    install_graphs1090
+fi
+
+if [[ "${run_tar1090_script}" == "true" ]]; then
+    install_tar1090
 fi
 
 exit 0